// src/app/types/data.type.ts
export type data = {
  name: string |null | undefined ;
  last_name: string |null | undefined ;
  phone: string |null | undefined ;
  country: string |null | undefined ;
  zip: string |null | undefined ;
  product: string |null | undefined ;
  address: string |null | undefined ;
  comment: string |null | undefined ;

}